--- components/example.ts
@@ -1,2 +1,0 @@
-// @ts-expect-error
-export { Container as default } from "./DropTarget.tsx";
