+++ package.json
@@ -19,6 +19,7 @@
     "@types/node": "18.8.2",
     "@types/react": "18.0.21",
     "@types/react-dom": "18.0.6",
+    "immutability-helper": "^3.1.1",
     "moment": "^2.29.4",
     "next": "12.3.1",
     "react": "18.2.0",
